test(js): keep #295 bundle-export test inline, use tempfile (Codacy)
The integration-test move dropped counted coverage 7.84% (integration-test coverage of
compile_project_with_jsx isn't attributed by the coverage tool). Restore the inline unit test — which
DOES count — but avoid Codacy's std::env::temp_dir() flag by using tempfile::tempdir() (RAII, unique
dir). Satisfies both the static-analysis and coverage-variation gates.
