+++ crates/tish_compile_js/src/tests_jsx.rs
@@ -167,6 +167,33 @@ mod tests {
         let _ = std::fs::remove_file(&path);
     }
 
+    /// #295: `--target js --format bundle` must re-emit the entry module's named exports as a real ES
+    /// `export { … }` (previously only `export default` surfaced; `export fn/const/{…}` were silently
+    /// dropped, forcing downstream regex post-processing of every bundle).
+    #[test]
+    fn bundle_emits_named_exports() {
+        // `tempfile::tempdir()` (RAII, unique dir) rather than a fixed `std::env::temp_dir()` path.
+        let dir = tempfile::tempdir().expect("tempdir");
+        let path = dir.path().join("entry.tish");
+        let src = "export fn mountFoo(h) { return h }\n\
+                   export const VERSION = \"1\"\n\
+                   fn helper() { return 2 }\n\
+                   export { helper as helperExported }\n\
+                   export default 42\n";
+        std::fs::write(&path, src).unwrap();
+        let js = compile_project_with_jsx(&path, Some(dir.path()), false)
+            .expect("compile_project_with_jsx failed");
+        assert!(
+            js.contains("export { VERSION, helper as helperExported, mountFoo };"),
+            "bundle must emit named exports, got:\n{}",
+            &js[js.len().saturating_sub(300)..]
+        );
+        assert!(
+            js.contains("export default __default_0;"),
+            "bundle must still emit the default export"
+        );
+    }
+
     #[test]
     fn jsx_never_emits_vdom_helpers_or_prelude_flags() {
         let src = r#"fn X() { return <div class="x">{"a"}</div> }"#;
